server/requirements.txt
bcrypt==4.2.0
chromadb==0.5.16
cryptography==43.0.3
fastapi==0.115.4
langchain==0.3.6
langchain-community==0.3.4
langchain-core==0.3.15
langchain-groq==0.2.1
langchain-nvidia-ai-endpoints==0.3.3
langchain-text-splitters==0.3.1
pypdf==5.1.0
python-jose==3.3.0
python-multipart==0.0.17
python-pptx==1.0.2
python-socketio==5.11.4
redis==5.2.0
unstructured==0.16.3
unstructured-client==0.26.2
unstructured-inference==0.8.1
unstructured-ingest==0.2.1
unstructured.pytesseract==0.3.13
uvicorn==0.32.0
websocket-client==1.8.0
websockets==13.1
aioredis
unstructured[docx]
unstructured[pptx]